content-security-policy/blink-contrib-2/base-uri-allow.sub.html

<!DOCTYPE html>
<html>

<head>
    <!-- Programmatically converted from a WebKit Reftest, please forgive resulting idiosyncrasies.-->
    <title>base-uri-allow</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src='../support/logTest.sub.js?logs=["TEST COMPLETE"]'></script>
    <!-- enforcing policy:
base-uri http://www1.{{host}}:{{ports[http][0]}}; script-src 'self' 'unsafe-inline' http://www1.{{host}}:{{ports[http][0]}}; connect-src 'self';
-->
    <base href="http://www1.{{host}}:{{ports[http][0]}}/">
    <script>
        'use strict';

        test(function() {
            const port = '{{ports[http][0]}}';
            if (port === '80' || port === '443') {
                assert_equals(document.baseURI, 'http://www1.{{host}}/');
            } else {
                assert_equals(document.baseURI, 'http://www1.{{host}}:' + port + '/');
            }

            log_test("TEST COMPLETE");
        });

    </script>
</head>

<body>
    <p>Check that base URIs can be set if they do not violate the page's policy.</p>
    <script async defer src="./content-security-policy/support/checkReport.sub.js?reportExists=false"></script>
</body>

</html>
